requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.3.26
langchain-community==0.3.27
openai==1.30.5
python-dotenv==1.0.1
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pytz==2024.1
streamlit==1.46.1
httpx==0.27.0
langchain-openai==0.3.27